requirements.txt
langextract>=0.1.0
python-dotenv>=1.0.1
pypdf>=5.0.0
